Reverted the full move of "maple" into "crates/" folder. Now there's "crates/maple_cli/lib.rs" with most code living there.
