+++ Cargo.toml
@@ -12,20 +12,7 @@ categories = ["command-line-utilities"]
 description = "vim-clap external filter gadget"
 
 [dependencies]
-regex = "1"
-rayon = "1.2"
-serde = { package = "serde", version = "1.0",  features = ["derive"] }
-anyhow = "1.0"
-structopt = "0.3"
-bytecount = "0.6"
-serde_json = "1.0"
-lazy_static = "1.4.0"
-fuzzy-matcher = "0.3.1"
-crossbeam-channel = "0.4"
-
-icon = { path = "../icon" }
-fuzzy_filter = { path = "../fuzzy_filter" }
-extracted_fzy = { path = "../extracted_fzy" }
+maple_cli = { path = "crates/maple_cli" }
 
 [build-dependencies]
 built = { package = "built", version = "0.4", features = ["git2"]}
