+++ crates/maple_cli/src/cmd/exec.rs
