+++ crates/maple_cli/src/cmd/helptags.rs
