+++ crates/maple_cli/src/cmd/mod.rs
