+++ crates/maple_cli/src/error.rs
