+++ crates/maple_cli/src/light_command.rs
